cleanup how servers stats work
