+++ obsidion/__main__.py
@@ -43,7 +43,6 @@ bot.load_extension("obsidion.cogs.misc")
 bot.load_extension("obsidion.cogs.rcon")
 bot.load_extension("obsidion.cogs.redstone")
 bot.load_extension("obsidion.cogs.servers")
-bot.load_extension("obsidion.cogs.hivestats")
 bot.load_extension("obsidion.cogs.servertracking")
 bot.load_extension("obsidion.cogs.events")
 bot.load_extension("obsidion.cogs.config")
