--- obsidion/cogs/hivestats/__init__.py
@@ -1,5 +1,0 @@
-from .hivestats import hivestats
-
-
-def setup(bot):
-    bot.add_cog(hivestats(bot))
