+++ obsidion/cogs/servers/servers.py
@@ -1,6 +1,17 @@
 import discord
 from discord.ext import commands
-from .utils import *
+from .utils import (
+    wyncraftClasses,
+    hiveMCStatus,
+    hiveMCGameStats,
+    hiveMCRank,
+    manacube,
+    blocksmc,
+    universocraft,
+    minesaga,
+    gommehd,
+    veltpvp,
+)
 from uuid import UUID
 from obsidion.utils.utils import usernameToUUID
 
@@ -270,4 +281,99 @@ class servers(commands.Cog):
             ),
         )
         await ctx.send(embed=embed)
+
+    @commands.command()
+    async def hiverank(self, ctx: commands.Context, username: str):
+        await ctx.trigger_typing()
+        data = await hiveMCRank(username, ctx.bot.http_session)
+        if not data:
+            await ctx.send(
+                f"`{username}` has not logged onto Hive or there are no ranks available."
+            )
+            return
+        embed = discord.Embed(color=0xFFAF03)
+        embed.set_author(
+            name=f"Hive rank for {username}",
+            url=f"https://www.hivemc.com/player/{username}",
+            icon_url="https://www.hivemc.com/img/white-logo.png",
+        )
+        embed.set_thumbnail(
+            url=f"https://visage.surgeplay.com/bust/{await usernameToUUID(username, ctx.bot.http_session)}"
+        )
+        embed.timestamp = ctx.message.created_at
+        embed.add_field(name="rank", value=(f"Rank: `{data['rank'][0]}`"))
+        await ctx.send(embed=embed)
+
+    @commands.command()
+    async def hivestatus(self, ctx: commands.Context, username: str):
+        await ctx.trigger_typing()
+        data = await hiveMCStatus(username, ctx.bot.http_session)
+        if not data:
+            await ctx.send(
+                f"`{username}` has not logged onto Hive or their status is not available."
+            )
+            return
+        embed = discord.Embed(color=0xFFAF03)
+        embed.set_author(
+            name=f"Hive Status for {username}",
+            url=f"https://www.hivemc.com/player/{username}",
+            icon_url="https://www.hivemc.com/img/white-logo.png",
+        )
+        embed.set_thumbnail(
+            url=f"https://visage.surgeplay.com/bust/{await usernameToUUID(username, ctx.bot.http_session)}"
+        )
+        embed.timestamp = ctx.message.created_at
+        embed.add_field(
+            name="description",
+            value=(f"Description: `{data['status'][0]['description']}`"),
+        )
+        embed.add_field(name="game", value=(f"Game: `{data['status'][0]['game']}`"))
+        await ctx.send(embed=embed)
+
+    @commands.command()
+    async def hivestats(self, ctx: commands.Context, username: str, game: str):
+        await ctx.trigger_typing()
+
+        if game.lower() in hive_con:
+            data = await hiveMCGameStats(
+                username, hive_con[game.lower()], ctx.bot.http_session
+            )
+            embed = discord.Embed(color=0xFFAF03)
+            embed.set_author(
+                name=f"Hive Stats for {username}",
+                url=f"https://www.hivemc.com/player/{username}",
+                icon_url="https://www.hivemc.com/img/white-logo.png",
+            )
+            embed.set_thumbnail(
+                url=f"https://visage.surgeplay.com/bust/{await usernameToUUID(username, ctx.bot.http_session)}"
+            )
+            embed.timestamp = ctx.message.created_at
+            if not data:
+                await ctx.send("No stats found")
+                return
+            del data["stats"][0]["UUID"]
+            if "cached" in data["stats"][0]:
+                del data["stats"][0]["cached"]
+            if "firstLogin" in data["stats"][0]:
+                del data["stats"][0]["firstLogin"]
+            if "lastLogin" in data["stats"][0]:
+                del data["stats"][0]["lastLogin"]
+            if "achievements" in data["stats"][0]:
+                del data["stats"][0]["achievements"]
+            if "title" in data["stats"][0]:
+                del data["stats"][0]["title"]
+            value = ""
+            for stat in data["stats"][0]:
+                if isinstance(data["stats"][0][stat], list) or isinstance(
+                    data["stats"][0][stat], dict
+                ):
+                    pass
+                else:
+                    value += f"`{stat}`: {data['stats'][0][stat]}\n"
+            embed.add_field(
+                name=f"{game.replace('_', ' ').upper()} Stats", value=value,
+            )
+            await ctx.send(embed=embed)
+        else:
+            await ctx.send("Sorry that game was not recognized as a Hive game")
